Minor fixes after iris update

+++ src/filesharingitem.cpp
@@ -97,7 +97,7 @@ FileSharingItem::FileSharingItem(const QImage &image, PsiAccount *acc, FileShari
 
     if (!initFromCache()) {
         _mimeType = QString::fromLatin1("image/png");
-        _fileSize = qint64(ba.size());
+        _fileSize = quint64(ba.size());
         QTemporaryFile file(QDir::tempPath() + QString::fromLatin1("/psishare-XXXXXX.png"));
         file.open();
         file.write(ba);
@@ -123,7 +123,7 @@ FileSharingItem::FileSharingItem(const QString &fileName, PsiAccount *acc, FileS
 
     if (!initFromCache()) {
         file.seek(0);
-        _fileSize = qint64(file.size());
+        _fileSize = quint64(file.size());
         _mimeType = QMimeDatabase().mimeTypeForFileNameAndData(fileName, &file).name();
     }
 }
@@ -138,7 +138,7 @@ FileSharingItem::FileSharingItem(const QString &mime, const QByteArray &data, co
 
     if (!initFromCache()) {
         _mimeType = mime;
-        _fileSize = qint64(data.size());
+        _fileSize = quint64(data.size());
 
         QMimeDatabase  db;
         QString        fileExt = db.mimeTypeForData(data).suffixes().value(0);
@@ -221,7 +221,7 @@ Reference FileSharingItem::toReference(const Jid &selfJid) const
     for (auto const &h : _sums)
         jfile.addHash(h);
     jfile.setName(fi.fileName());
-    jfile.setSize(fi.size());
+    jfile.setSize(quint64(fi.size()));
     jfile.setMediaType(_mimeType);
     jfile.setDescription(_description);
 
@@ -373,7 +373,7 @@ void FileSharingItem::publish(const XMPP::Jid &myJid)
     }
 }
 
-FileShareDownloader *FileSharingItem::download(bool isRanged, qint64 start, qint64 size)
+FileShareDownloader *FileSharingItem::download(bool isRanged, qint64 start, quint64 size)
 {
     if (isRanged && (_flags & SizeKnown) && start == 0 && size == _fileSize)
         isRanged = false;
